Tests passing

+++ src/main/kotlin/com/grondona/model/dto/UserGroupResponse.kt
@@ -7,7 +7,7 @@ import java.util.UUID
 data class UserGroupResponse(
     val groupId: UUID,
     val name: String,
-    val memberCount: Int,
+    val memberCount: Long,
     val points: Float = 0f,
     val rank: Int? = null,
     val role: GroupRole = GroupRole.MEMBER,
@@ -15,7 +15,7 @@ data class UserGroupResponse(
     constructor(
         groupId: UUID,
         name: String,
-        memberCount: Int,
+        memberCount: Long,
         points: Float,
         role: GroupRole = GroupRole.MEMBER,
     ) : this(groupId, name, memberCount, points, null, role)
